requirements.txt
appdirs==1.4.4
asgiref==3.5.2
audioread==3.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
fontawesomefree==6.1.2
gunicorn==20.1.0
idna==3.3
joblib==1.2.0
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pooch==1.6.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
scikit-learn==1.1.2
scipy==1.9.1
SoundFile==0.10.3.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
